### 2020-10-09 #2737 @whedon
[PRE REVIEW]: pystiche: A Framework for Neural Style Transfer
**Submitting author:** @pmeier (<a href="http://orcid.org/0000-0002-5184-1622">Philip Meier</a>)
**Repository:** <a href="https://github.com/pmeier/pystiche" target ="_blank">https://github.com/pmeier/pystiche</a>
**Version:** v0.6.1
