### 2020-10-09 #2737 @kthyng
@pmeier Also I've read through your paper several times but I'm having a hard time figuring out a few categories of where this software fits in order to match it up with our editors. You suggested arfon as editor when you submitted but that doesn't seem like a good match to me. We don't have someone who does image processing in particular, but some who do machine learning in particular domains. Does this work fit into a particular domain?
