### 2020-10-09 #2737 @pmeier
Hey @kthyng

> Also I've read through your paper several times but I'm having a hard time figuring out a few categories of where this software fits in order to match it up with our editors. You suggested arfon as editor when you submitted but that doesn't seem like a good match to me.

`pystiche` was already reviewed in pyOpenSci/software-review#25 in a previous step. The editor @NickleDave ["tasked" me](https://github.com/pyOpenSci/software-review/issues/25#issuecomment-705646440) to submit to JOSS. In the same comment he tagged @arfon, hence my suggestion.

> We don't have someone who does image processing in particular, but some who do machine learning in particular domains. Does this work fit into a particular domain?

Yes, machine learning plays a big role in the project. Looking through your [list of editors](https://joss.theoj.org/about#editorial_board) I think Juan Nunez-Iglesias (jni) would also be a good fit since he is listed for Python and image processing.
