### 2020-10-10 #2737 @kthyng
Ok so @pmeier let me know when you have added a statement of need and I think we can proceed through the much shortened process since this came from pyopensci.
