### 2020-10-12 pmeier/pystiche#422 @pmeier
Missing DOIs in literature references
> ```
> MISSING DOIs
>
> - 10.1109/cvpr.2015.7299155 may be a valid DOI for title: Understanding Deep Image Representations by Inverting Them
> - 10.1007/springerreference_62539 may be a valid DOI for title: Non-Photorealistic Rendering
> ```

The first one is valid while the second seems to be invalid. At least the DOI [doesn't resolve properly](https://www.doi.org/10.1007/springerreference_62539) and I can't find another DOI manually.
