+++ paper.md
@@ -16,7 +16,7 @@ affiliations:
  - name: inIT –- Institute Industrial IT, Technische Hochschule Ostwestfalen-Lippe (TH-OWL)
    index: 1
    
-date: 2 July 2020
+date: 15 October 2020
 bibliography: docs/source/references.bib
 ---
 
