### 2020-10-15 pmeier/pystiche#424 @pmeier
JOSS Paper is missing a "Statement of Need" section
Currently the JOSS paper only includes a ["Summary & Purpose"](https://github.com/pmeier/pystiche/blame/3202809dd6dcf03c1d65040198d4a69c1a55a0c2/paper.md#L23) section as mentioned in https://github.com/openjournals/joss-reviews/issues/2737#issuecomment-706221452.

The [guidelines](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) state:

> - A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.
> - A clear Statement of Need that illustrates the research purpose of the software.

Thus, we need to add a "Statement of Need" section.
