### 2020-10-15 3dc4bbb
add arxiv as prefix to other references as well
