+++ docs/source/references.bib
@@ -1,13 +1,14 @@
 % Encoding: UTF-8
 
 @InProceedings{GEB2016,
-  author    = {Gatys, Leon A. and Ecker, Alexander. S. and Bethge, Matthias},
-  booktitle = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
-  title     = {Image Style Transfer Using Convolutional Neural Networks},
-  year      = {2016},
-  abstract  = {Rendering the semantic content of an image in different styles is a difficult image processing task. Arguably, a major limiting factor for previous approaches has been the lack of image representations that explicitly represent semantic information and, thus, allow to separate image content from style. Here we use image representations derived from Convolutional Neural Networks optimised for object recognition, which make high level image information explicit. We introduce A Neural Algorithm of Artistic Style that can separate and recombine the image content and style of natural images. The algorithm allows us to produce new images of high perceptual quality that combine the content of an arbitrary photograph with the appearance of numerous wellknown artworks. Our results provide new insights into the deep image representations learned by Convolutional Neural Networks and demonstrate their potential for high level image synthesis and manipulation.},
-  doi       = {10.1109/CVPR.2016.265},
-  eprint    = {1508.06576},
+  author        = {Gatys, Leon A. and Ecker, Alexander. S. and Bethge, Matthias},
+  booktitle     = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
+  title         = {Image Style Transfer Using Convolutional Neural Networks},
+  year          = {2016},
+  abstract      = {Rendering the semantic content of an image in different styles is a difficult image processing task. Arguably, a major limiting factor for previous approaches has been the lack of image representations that explicitly represent semantic information and, thus, allow to separate image content from style. Here we use image representations derived from Convolutional Neural Networks optimised for object recognition, which make high level image information explicit. We introduce A Neural Algorithm of Artistic Style that can separate and recombine the image content and style of natural images. The algorithm allows us to produce new images of high perceptual quality that combine the content of an arbitrary photograph with the appearance of numerous wellknown artworks. Our results provide new insights into the deep image representations learned by Convolutional Neural Networks and demonstrate their potential for high level image synthesis and manipulation.},
+  archiveprefix = {arXiv},
+  doi           = {10.1109/CVPR.2016.265},
+  eprint        = {1508.06576},
 }
 
 @Article{SZ2014,
@@ -39,17 +40,19 @@
   year      = {2016},
   abstract  = {This paper studies a combination of generative Markov random field (MRF) models and discriminatively trained deep convolutional neural networks (dCNNs) for synthesizing 2D images. The generative MRF acts on higher-levels of a dCNN feature pyramid, controlling the image layout at an abstract level. We apply the method to both photographic and non-photo-realistic (artwork) synthesis tasks. The MRF regularizer prevents over-excitation artifacts and reduces implausible feature mixtures common to previous dCNN inversion approaches, permitting synthesizing photographic content with increased visual plausibility. Unlike standard MRF-based texture synthesis, the combined system can both match and adapt local features with considerable variability, yielding results far out of reach of classic generative MRF methods.},
   doi       = {10.1109/CVPR.2016.272},
+  archiveprefix = {arXiv},
   eprint    = {1601.04589},
 }
 
 @InProceedings{GEB+2017,
-  author    = {Gatys, Leon A. and Ecker, Alexander S. and Bethge, Matthias and Hertzmann, Aaron and Shechtman, Eli},
-  booktitle = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
-  title     = {Controlling Perceptual Factors in Neural Style Transfer},
[…]
