+++ docs/source/references.bib
@@ -16,7 +16,7 @@
   journal       = {Computing Research Repository ({CoRR})},
   title         = {Very Deep Convolutional Networks for Large-Scale Image Recognition},
   year          = {2014},
-  volume        = {abs/1409.1556},
+  volume        = {abs/1409},
   abstract      = {In this work we investigate the effect of the convolutional network depth on its accuracy in the large-scale image recognition setting. Our main contribution is a thorough evaluation of networks of increasing depth using an architecture with very small (3x3) convolution filters, which shows that a significant improvement on the prior-art configurations can be achieved by pushing the depth to 16-19 weight layers. These findings were the basis of our ImageNet Challenge 2014 submission, where our team secured the first and the second places in the localisation and classification tracks respectively. We also show that our representations generalise well to other datasets, where they achieve state-of-the-art results. We have made our two best-performing ConvNet models publicly available to facilitate further research on the use of deep visual representations in computer vision.},
   archiveprefix = {arXiv},
   eprint        = {1409.1556},
