+++ docs/source/references.bib
@@ -20,15 +20,15 @@
   eprint   = {1409.1556},
 }
 
-@Article{MV2014,
-  author   = {Mahendran, Aravindh and Vedaldi, Andrea},
-  journal  = {Computing Research Repository ({CoRR})},
-  title    = {Understanding Deep Image Representations by Inverting Them},
-  year     = {2014},
-  volume   = {abs/1412.0035},
-  abstract = {Image representations, from SIFT and Bag of Visual Words to Convolutional Neural Networks (CNNs), are a crucial component of almost any image understanding system. Nevertheless, our understanding of them remains limited. In this paper we conduct a direct analysis of the visual information contained in representations by asking the following question: given an encoding of an image, to which extent is it possible to reconstruct the image itself? To answer this question we contribute a general framework to invert representations. We show that this method can invert representations such as HOG and SIFT more accurately than recent alternatives while being applicable to CNNs too. We then use this technique to study the inverse of recent state-of-the-art CNN image representations for the first time. Among our findings, we show that several layers in CNNs retain photographically accurate information about the image, with different degrees of geometric and photometric invariance.},
-  doi      = {10.1109/CVPR.2015.7299155},
-  eprint   = {1412.0035},
+@InProceedings{MV2015,
+  author        = {Mahendran, Aravindh and Vedaldi, Andrea},
+  booktitle     = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
+  title         = {Understanding Deep Image Representations by Inverting Them},
+  year          = {2015},
+  abstract      = {Image representations, from SIFT and Bag of Visual Words to Convolutional Neural Networks (CNNs), are a crucial component of almost any image understanding system. Nevertheless, our understanding of them remains limited. In this paper we conduct a direct analysis of the visual information contained in representations by asking the following question: given an encoding of an image, to which extent is it possible to reconstruct the image itself? To answer this question we contribute a general framework to invert representations. We show that this method can invert representations such as HOG and SIFT more accurately than recent alternatives while being applicable to CNNs too. We then use this technique to study the inverse of recent state-of-the-art CNN image representations for the first time. Among our findings, we show that several layers in CNNs retain photographically accurate information about the image, with different degrees of geometric and photometric invariance.},
+  doi           = {10.1109/CVPR.2015.7299155},
+  archiveprefix = {arXiv},
+  eprint        = {1412.0035},
 }
 
 @InProceedings{LW2016,
