+++ docs/source/references.bib
@@ -34,14 +34,14 @@
 }
 
 @InProceedings{LW2016,
-  author    = {Li, Chuan and Wand, Michael},
-  booktitle = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
-  title     = {Combining Markov Random Fields and Convolutional Neural Networks for Image Synthesis},
-  year      = {2016},
-  abstract  = {This paper studies a combination of generative Markov random field (MRF) models and discriminatively trained deep convolutional neural networks (dCNNs) for synthesizing 2D images. The generative MRF acts on higher-levels of a dCNN feature pyramid, controlling the image layout at an abstract level. We apply the method to both photographic and non-photo-realistic (artwork) synthesis tasks. The MRF regularizer prevents over-excitation artifacts and reduces implausible feature mixtures common to previous dCNN inversion approaches, permitting synthesizing photographic content with increased visual plausibility. Unlike standard MRF-based texture synthesis, the combined system can both match and adapt local features with considerable variability, yielding results far out of reach of classic generative MRF methods.},
-  doi       = {10.1109/CVPR.2016.272},
+  author        = {Li, Chuan and Wand, Michael},
+  booktitle     = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
+  title         = {Combining Markov Random Fields and Convolutional Neural Networks for Image Synthesis},
+  year          = {2016},
+  abstract      = {This paper studies a combination of generative Markov random field (MRF) models and discriminatively trained deep convolutional neural networks (dCNNs) for synthesizing 2D images. The generative MRF acts on higher-levels of a dCNN feature pyramid, controlling the image layout at an abstract level. We apply the method to both photographic and non-photo-realistic (artwork) synthesis tasks. The MRF regularizer prevents over-excitation artifacts and reduces implausible feature mixtures common to previous dCNN inversion approaches, permitting synthesizing photographic content with increased visual plausibility. Unlike standard MRF-based texture synthesis, the combined system can both match and adapt local features with considerable variability, yielding results far out of reach of classic generative MRF methods.},
+  doi           = {10.1109/CVPR.2016.272},
   archiveprefix = {arXiv},
-  eprint    = {1601.04589},
+  eprint        = {1601.04589},
 }
 
 @InProceedings{GEB+2017,
