### 2020-10-15 1922681
Add a Statement of Need to the JOSS Paper (#425)
* split into summary and statement of need

* fix date

* fix references

* try to enable arxiv link

* add arxiv as prefix to other references as well

* try folded style

* remove paper number from CoRR volumes

* revise paper

* fix reference indentation
