+++ paper.md
@@ -13,14 +13,16 @@ authors:
     orcid: 0000-0002-3325-7887
     affiliation: 1
 affiliations:
- - name: inIT –- Institute Industrial IT, Technische Hochschule Ostwestfalen-Lippe (TH-OWL)
+ - name: >
+     inIT –- Institute Industrial IT, 
+     Technische Hochschule Ostwestfalen-Lippe (TH-OWL)
    index: 1
    
-date: 2 July 2020
+date: 15 October 2020
 bibliography: docs/source/references.bib
 ---
 
-# Summary and Purpose
+# Summary
 
 The seminal work of Gatys, Ecker, and Bethge gave birth to the field of 
 _Neural Style Transfer_ (NST) in 2016 [@GEB2016]. The general idea behind an NST can be 
@@ -29,19 +31,37 @@ conveyed with only three images and two symbols:
 ![](docs/source/graphics/banner/banner.jpg)
 
 In words: an NST describes the merger between the content and artistic style of two 
-arbitrary images. This idea is nothing new in the field of computational style transfer 
-[@GG2001]. What distinguishes NST from traditional approaches is its generality: an NST 
-only needs a single arbitrary content and style image as input and thus "makes -- for 
-the first time -- a generalized style transfer practicable" [@SID2017].
+arbitrary images. This idea is nothing new in the field of 
+_Non-photorealistic rendering_ (NPR) [@GG2001]. What distinguishes NST from traditional 
+NPR approaches is its generality: an NST only needs a single arbitrary content and 
+style image as input and thus "makes -- for the first time -- a generalized style 
+transfer practicable" [@SID2017].
 
-Due to its vivid nature, the field of NST gained a lot of traction after its emergence 
-[@JYF+2019]. While many new techniques or augmentations have been developed, the field 
-lacks standardization, which is especially evident in the reference implementations of 
-the authors. `pystiche` aims to fill this gap.
+Besides peripheral tasks, an NST at its core is the definition of an optimization 
+criterion called _perceptual loss_, which estimates the perceptual quality of the 
+stylized image. Usually the perceptual loss comprises a deep neural network that needs 
+to supply encodings of images from various depths.
 
-`pystiche` is a framework for NST written in Python. It is built upon the Deep Learning 
-(DL) framework PyTorch [@PGM+2019], since at the core each NST algorithm utilizes a 
-Deep Neural Network. `pystiche` pursues similar goals as DL frameworks:
+`pystiche` is a framework for NST written in Python and built upon the _Deep Learning_ 
+(DL) framework PyTorch [@PGM+2019]. It provides modular and efficient implementations 
+for commonly used perceptual losses [@MV2015; @GEB2016; @LW2016] as well as neural net 
+architectures [@SZ2014; @KSH2012]. This enables users to mix current state-of-the-art 
+techniques with new ideas with ease. 
+
+Due to its vivid nature, the field of NST gained a lot of traction in the short time 
+after its emergence [@JYF+2019]. While many new techniques or augmentations have been 
+developed, the field lacks standardization, which is especially evident in the 
+reference implementations of the authors. `pystiche` aims to fill this gap.
+
+# Statement of Need
+
+Currently, unlike DL, there exist no library or framework for implementing NST. Thus, 
+authors of new NST techniques either implement everything from scratch or base their 
+implementation upon existing ones of other authors. Both ways have their downsides: 
+while the former dampens innovation due to the lengthy implementation of reusable 
+parts, with the latter the author inherits the technical debt due to the rapid 
+development pace of DL hard- and software. In order to overcome this, `pystiche` 
+pursues similar goals as DL frameworks:
 
 1. **Accessibility**
[…]
