+++ docs/source/references.bib
@@ -1,23 +1,25 @@
 % Encoding: UTF-8
 
 @InProceedings{GEB2016,
-  author    = {Gatys, Leon A. and Ecker, Alexander. S. and Bethge, Matthias},
-  booktitle = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
-  title     = {Image Style Transfer Using Convolutional Neural Networks},
-  year      = {2016},
-  abstract  = {Rendering the semantic content of an image in different styles is a difficult image processing task. Arguably, a major limiting factor for previous approaches has been the lack of image representations that explicitly represent semantic information and, thus, allow to separate image content from style. Here we use image representations derived from Convolutional Neural Networks optimised for object recognition, which make high level image information explicit. We introduce A Neural Algorithm of Artistic Style that can separate and recombine the image content and style of natural images. The algorithm allows us to produce new images of high perceptual quality that combine the content of an arbitrary photograph with the appearance of numerous wellknown artworks. Our results provide new insights into the deep image representations learned by Convolutional Neural Networks and demonstrate their potential for high level image synthesis and manipulation.},
-  doi       = {10.1109/CVPR.2016.265},
-  eprint    = {1508.06576},
+  author        = {Gatys, Leon A. and Ecker, Alexander. S. and Bethge, Matthias},
+  booktitle     = {{IEEE} Conference on Computer Vision and Pattern Recognition ({CVPR})},
+  title         = {Image Style Transfer Using Convolutional Neural Networks},
+  year          = {2016},
+  abstract      = {Rendering the semantic content of an image in different styles is a difficult image processing task. Arguably, a major limiting factor for previous approaches has been the lack of image representations that explicitly represent semantic information and, thus, allow to separate image content from style. Here we use image representations derived from Convolutional Neural Networks optimised for object recognition, which make high level image information explicit. We introduce A Neural Algorithm of Artistic Style that can separate and recombine the image content and style of natural images. The algorithm allows us to produce new images of high perceptual quality that combine the content of an arbitrary photograph with the appearance of numerous wellknown artworks. Our results provide new insights into the deep image representations learned by Convolutional Neural Networks and demonstrate their potential for high level image synthesis and manipulation.},
+  archiveprefix = {arXiv},
+  doi           = {10.1109/CVPR.2016.265},
+  eprint        = {1508.06576},
 }
 
 @Article{SZ2014,
-  author   = {Simonyan, Karen and Zisserman, Andrew},
-  journal  = {Computing Research Repository ({CoRR})},
-  title    = {Very Deep Convolutional Networks for Large-Scale Image Recognition},
-  year     = {2014},
-  volume   = {abs/1409.1556},
-  abstract = {In this work we investigate the effect of the convolutional network depth on its accuracy in the large-scale image recognition setting. Our main contribution is a thorough evaluation of networks of increasing depth using an architecture with very small (3x3) convolution filters, which shows that a significant improvement on the prior-art configurations can be achieved by pushing the depth to 16-19 weight layers. These findings were the basis of our ImageNet Challenge 2014 submission, where our team secured the first and the second places in the localisation and classification tracks respectively. We also show that our representations generalise well to other datasets, where they achieve state-of-the-art results. We have made our two best-performing ConvNet models publicly available to facilitate further research on the use of deep visual representations in computer vision.},
-  eprint   = {1409.1556},
+  author        = {Simonyan, Karen and Zisserman, Andrew},
[…]
