### 2020-10-19 #2737 @kthyng
I want to be clear to link to the [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/25) already completed.
