### 2020-10-19 #2761 @kthyng
Reference completed [pyOpenSci review](https://github.com/pyOpenSci/software-review/issues/25).
