### 2020-10-19 #2761 @kthyng
@pmeier Thanks! At JOSS we ask for the title and author list in the metadata for the zenodo archive to exactly match your JOSS paper, though it isn't required. Could you do this?
