### 2020-10-19 #2761 @pmeier
Will try. The Zenodo archive is generated automatically through a webhook after a GitHub release. I'm not sure if I'm allowed to modify the meta data.
