### 2020-10-19 #2761 @pmeier
@kthyng I've just released [`v0.6.2`](https://github.com/pmeier/pystiche/releases/tag/v0.6.2) which fixes the metadata.
