### 2020-10-19 #2761 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
