### 2020-10-19 #2761 @kthyng
@pmeier congrats on your publication with JOSS! I'll close this issue once the doi resolves.
